Rename program-financial-service-provider-configurations folder
