+++ interfaces/portal/src/app/domains/financial-service-provider-configuration/financial-service-provider-configuration.model.ts
@@ -1,4 +1,4 @@
-import { ProgramFinancialServiceProviderConfigurationResponseDto } from '@121-service/src/program-financial-service-provider-configurations/dtos/program-financial-service-provider-configuration-response.dto';
+import { ProgramFinancialServiceProviderConfigurationResponseDto } from '@121-service/src/program-fsp-configurations/dtos/program-financial-service-provider-configuration-response.dto';
 
 import { Dto } from '~/utils/dto-type';
