--- services/121-service/src/program-financial-service-provider-configurations/program-financial-service-provider-configurations.module.ts
@@ -1,27 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-
-import { ProgramFinancialServiceProviderConfigurationEntity } from '@121-service/src/program-financial-service-provider-configurations/entities/program-financial-service-provider-configuration.entity';
-import { ProgramFinancialServiceProviderConfigurationPropertyEntity } from '@121-service/src/program-financial-service-provider-configurations/entities/program-financial-service-provider-configuration-property.entity';
-import { ProgramFinancialServiceProviderConfigurationsController } from '@121-service/src/program-financial-service-provider-configurations/program-financial-service-provider-configurations.controller';
-import { ProgramFinancialServiceProviderConfigurationRepository } from '@121-service/src/program-financial-service-provider-configurations/program-financial-service-provider-configurations.repository';
-import { ProgramFinancialServiceProviderConfigurationsService } from '@121-service/src/program-financial-service-provider-configurations/program-financial-service-provider-configurations.service';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([
-      ProgramFinancialServiceProviderConfigurationEntity,
-      ProgramFinancialServiceProviderConfigurationPropertyEntity,
-    ]),
-  ],
-  providers: [
-    ProgramFinancialServiceProviderConfigurationsService,
-    ProgramFinancialServiceProviderConfigurationRepository,
-  ],
-  controllers: [ProgramFinancialServiceProviderConfigurationsController],
-  exports: [
-    ProgramFinancialServiceProviderConfigurationRepository,
-    ProgramFinancialServiceProviderConfigurationsService,
-  ],
-})
-export class ProgramFinancialServiceProviderConfigurationsModule {}
