+++ services/121-service/src/program-fsp-configurations/dtos/update-program-financial-service-provider-configuration.dto.ts
@@ -8,7 +8,7 @@ import {
   ValidateNested,
 } from 'class-validator';
 
-import { CreateProgramFinancialServiceProviderConfigurationPropertyDto } from '@121-service/src/program-financial-service-provider-configurations/dtos/create-program-financial-service-provider-configuration-property.dto';
+import { CreateProgramFinancialServiceProviderConfigurationPropertyDto } from '@121-service/src/program-fsp-configurations/dtos/create-program-financial-service-provider-configuration-property.dto';
 import { LocalizedString } from '@121-service/src/shared/types/localized-string.type';
 
 export class UpdateProgramFinancialServiceProviderConfigurationDto {
