+++ services/121-service/src/program-fsp-configurations/entities/program-financial-service-provider-configuration-property.entity.ts
@@ -10,7 +10,7 @@ import {
 
 import { Base121Entity } from '@121-service/src/base.entity';
 import { FinancialServiceProviderConfigurationProperties } from '@121-service/src/fsps/enums/fsp-name.enum';
-import { ProgramFinancialServiceProviderConfigurationEntity } from '@121-service/src/program-financial-service-provider-configurations/entities/program-financial-service-provider-configuration.entity';
+import { ProgramFinancialServiceProviderConfigurationEntity } from '@121-service/src/program-fsp-configurations/entities/program-financial-service-provider-configuration.entity';
 
 @Unique('programFinancialServiceProviderConfigurationPropertyUnique', [
   'programFinancialServiceProviderConfigurationId',
