+++ services/121-service/src/program-fsp-configurations/entities/program-financial-service-provider-configuration.entity.ts
@@ -11,7 +11,7 @@ import {
 import { Base121Entity } from '@121-service/src/base.entity';
 import { FinancialServiceProviders } from '@121-service/src/fsps/enums/fsp-name.enum';
 import { TransactionEntity } from '@121-service/src/payments/transactions/transaction.entity';
-import { ProgramFinancialServiceProviderConfigurationPropertyEntity } from '@121-service/src/program-financial-service-provider-configurations/entities/program-financial-service-provider-configuration-property.entity';
+import { ProgramFinancialServiceProviderConfigurationPropertyEntity } from '@121-service/src/program-fsp-configurations/entities/program-financial-service-provider-configuration-property.entity';
 import { ProgramEntity } from '@121-service/src/programs/program.entity';
 import { RegistrationEntity } from '@121-service/src/registration/registration.entity';
 import { LocalizedString } from '@121-service/src/shared/types/localized-string.type';
